KA10: Removed compiler warnings.

+++ PDP10/ka10_imx.c
@@ -207,12 +207,12 @@ t_stat imx_devio(uint32 dev, uint64 *data)
 
 t_stat imx_svc (UNIT *uptr)
 {
-    int max_samples;
+    uint64 max_samples;
 
     if (status & IMX_PACK) {
-        max_samples = 3;
+        max_samples = 3LL;
     } else {
-        max_samples = 1;
+        max_samples = 1LL;
     }
 
     if (imx_samples < max_samples) {
