+++ PDP10/kx10_imp.c
@@ -405,7 +405,7 @@ struct imp_device {
     struct imp_map    port_map[64];            /* Ports to adjust */
     in_addr_T         dhcpip;                  /* DHCP server address */
     uint8             dhcp_state;              /* State of DHCP */
-    int               dhcp_lease;              /* DHCP lease time */
+    uint32            dhcp_lease;              /* DHCP lease time */
     int               dhcp_renew;              /* DHCP renew time */
     int               dhcp_rebind;             /* DHCP rebind time */
     int               dhcp_wait_time;          /* seconds waiting for response */
@@ -2639,7 +2639,7 @@ t_stat imp_reset (DEVICE *dptr)
     imp_data.sendq = NULL;
     /* Set up free queue */
     p = NULL;
-    for (i = 0; i < (sizeof(imp_buffer)/sizeof(struct imp_packet)); i++) {
+    for (i = 0; i < (int)(sizeof(imp_buffer)/sizeof(struct imp_packet)); i++) {
         imp_buffer[i].next = p;
         p = &imp_buffer[i];
     }
